Devirtualize the table slice builder
It's been almost a year since we've had more than one implementation of
the table slice builder. We cannot really justify having every call to
`table_slice_builder::add(...)` for every individual cell of data added
to VAST go through virtual dispatch. This simply removes the pure
virtual base class, renames `arrow_table_slice_builder` to
`table_slice_builder`, and removes the `factory<table_slice_builder>`
specialization.

In very much non-scientific tests this has increased the JSON reader
throughput by 2-3% on my local machine. I didn't bother to benchmark any
further, as we have quite a few performance-related changes for the next
release that we can benchmark togethet to save on time. Ultimately,
that's gonna be the numbers that's actually interesting to users.
