+++ libvast/builtins/commands/rebuild.cpp
@@ -7,7 +7,6 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 #include <vast/arrow_table_slice.hpp>
-#include <vast/arrow_table_slice_builder.hpp>
 #include <vast/concept/parseable/to.hpp>
 #include <vast/concept/parseable/vast/expression.hpp>
 #include <vast/data.hpp>
@@ -29,6 +28,7 @@
 #include <vast/system/spawn_or_connect_to_node.hpp>
 #include <vast/system/status.hpp>
 #include <vast/table_slice.hpp>
+#include <vast/table_slice_builder.hpp>
 #include <vast/uuid.hpp>
 
 #include <arrow/table.h>
