+++ libvast/builtins/operators/identity.cpp
@@ -10,7 +10,6 @@
 #include <vast/logger.hpp>
 #include <vast/pipeline.hpp>
 #include <vast/plugin.hpp>
-#include <vast/table_slice_builder_factory.hpp>
 
 #include <arrow/type.h>
