+++ libvast/builtins/operators/select.cpp
@@ -12,7 +12,6 @@
 #include <vast/error.hpp>
 #include <vast/pipeline.hpp>
 #include <vast/plugin.hpp>
-#include <vast/table_slice_builder_factory.hpp>
 #include <vast/type.hpp>
 
 #include <arrow/type.h>
