+++ libvast/include/vast/table_slice_builder.hpp
@@ -3,7 +3,7 @@
 //   | |/ / __ |_\ \  / /          Across
 //   |___/_/ |_/___/ /_/       Space and Time
 //
-// SPDX-FileCopyrightText: (c) 2018 The VAST Contributors
+// SPDX-FileCopyrightText: (c) 2019 The VAST Contributors
 // SPDX-License-Identifier: BSD-3-Clause
 
 #pragma once
@@ -11,44 +11,34 @@
 #include "vast/fwd.hpp"
 
 #include "vast/table_slice.hpp"
-#include "vast/type.hpp"
-#include "vast/view.hpp"
 
-#include <caf/ref_counted.hpp>
+#include <arrow/type.h>
+#include <flatbuffers/flatbuffers.h>
 
+#include <memory>
 #include <span>
-#include <type_traits>
+#include <vector>
 
 namespace vast {
 
-/// Enables incremental construction of a table slice.
-/// @relates table_slice
-class table_slice_builder : public caf::ref_counted {
+/// A builder for table slices that store elements encoded in the
+/// [Arrow](https://arrow.apache.org) format.
+class table_slice_builder final {
 public:
-  // -- types and constants ----------------------------------------------------
-
   /// The default size of the buffer that the builder works with.
   static constexpr size_t default_buffer_size = 8192;
 
   // -- constructors, destructors, and assignment operators --------------------
 
-  /// Forbid default-construction.
-  table_slice_builder() = delete;
+  /// Constructs an Arrow table slice builder instance.
+  /// @param layout The layout of the slice.
+  /// @param initial_buffer_size The buffer size the builder starts with.
+  /// @returns A table_slice_builder instance.
+  explicit table_slice_builder(type layout, size_t initial_buffer_size
+                                            = default_buffer_size);
 
-  /// Forbid copy-construction.
-  table_slice_builder(const table_slice_builder&) = delete;
-
-  /// Forbid copy-assignment.
-  table_slice_builder& operator=(const table_slice_builder&) = delete;
-
-  /// Forbid move-construction.
-  table_slice_builder(table_slice_builder&&) = delete;
-
-  /// Forbid move-assignment.
-  table_slice_builder& operator=(table_slice_builder&&) = delete;
-
-  /// Destroys a table slice builder.
-  virtual ~table_slice_builder() noexcept;
+  /// Destroys an Arrow table slice builder.
+  ~table_slice_builder() noexcept;
 
   // -- properties -------------------------------------------------------------
 
@@ -64,73 +54,164 @@ public:
   [[nodiscard]] bool add(const T& x, const Ts&... xs) {
     if constexpr (sizeof...(xs) == 0) {
       if constexpr (std::is_same_v<T, data_view>) {
-        return add_impl(x);
+        return add(x);
       } else if constexpr (caf::detail::tl_contains<data_view::types, T>::value) {
-        return add_impl(x);
+        return add(data_view{x});
       } else {
-        return add_impl(make_view(x));
+        return add(make_view(x));
       }
     } else {
       return add(x) && (add(xs) && ...);
     }
   }
 
-  /// Constructs a table_slice from the currently accumulated state. After
-  /// calling this function, implementations must reset their internal state
-  /// such that subsequent calls to add will restart with a new table_slice.
-  /// @returns A table slice from the accumulated calls to add.
-  /// @note Returns an invalid table slice on failure.
-  [[nodiscard]] virtual table_slice finish() = 0;
+  /// Adds data to the builder.
+  /// @param x The data to add.
+  /// @returns `true` on success.
+  bool add(data_view x);
 
-  /// @returns The current number of rows in the table slice.
-  virtual size_t rows() const noexcept = 0;
+  [[nodiscard]] table_slice finish();
+
+  /// Creates a table slice from a record batch.
+  /// @pre `record_batch->schema()->Equals(make_experimental_schema(layout))``
+  /// @param batch A pre-existing record batch.
+  /// @param schema VAST schema matching the record batch schema. Parameter
+  ///     is optional and derived from the record batch if not provided.
+  [[nodiscard]] table_slice static create(
+    const std::shared_ptr<arrow::RecordBatch>& record_batch, type schema = {},
+    table_slice::serialize serialize = table_slice::serialize::no,
+    size_t initial_buffer_size = default_buffer_size);
 
   /// @returns The number of columns in the table slice.
   size_t columns() const noexcept;
 
-  /// @returns The table layout.
-  const type& layout() const noexcept;
-
-  /// @returns An identifier for the implementing class.
-  virtual table_slice_encoding implementation_id() const noexcept = 0;
+  /// @returns The current number of rows in the table slice.
+  size_t rows() const noexcept;
 
   /// Allows The table slice builder to allocate sufficient storage.
   /// @param `num_rows` The number of rows to allocate storage for.
-  /// @note The default implementation does nothing.
-  virtual void reserve(size_t num_rows);
+  void reserve(size_t num_rows);
 
-  /// Opt-in to CAF's type inspection API.
-  template <class Inspector>
-  friend auto inspect(Inspector& f, table_slice_builder& x) {
-    return f.object(x)
-      .pretty_name("vast.table_slice_builder")
-      .fields(f.field("layout", x.layout_));
-  }
-
-protected:
-  // -- implementation utilities -----------------------------------------------
-
-  /// Construct a builder for tables slices.
-  /// @param layout The table layout.
-  explicit table_slice_builder(type layout) noexcept;
-
-  /// Adds data to the builder.
-  /// @param x The data to add.
-  /// @returns `true` on success.
-  virtual bool add_impl(data_view x) = 0;
+  /// @returns The table layout.
+  const type& layout() const noexcept;
 
 private:
   // -- implementation details -------------------------------------------------
 
+  // The VAST schema this builder was created from.
   type layout_;
+
+  /// A flattened representation of the schema that is iterated over when
+  /// calling add.
+  std::vector<record_type::leaf_view> leaves_;
+  std::vector<record_type::leaf_view>::iterator current_leaf_;
+
+  /// Number of filled rows.
+  size_t num_rows_ = 0;
+
+  /// The serialized layout can be cached because every builder instance only
+  /// produces slices of a single layout.
+  mutable std::vector<char> serialized_layout_cache_;
+
+  /// Schema of the Record Batch corresponding to the layout.
+  std::shared_ptr<arrow::Schema> schema_ = {};
+
+  /// Underlying Arrow builder for record batches.
+  std::shared_ptr<arrow::ArrayBuilder> arrow_builder_;
+
+  /// The underlying FlatBuffers builder.
+  flatbuffers::FlatBufferBuilder builder_;
 };
 
-// -- intrusive_ptr facade -----------------------------------------------------
+using table_slice_builder_ptr = std::shared_ptr<table_slice_builder>;
 
-/// @relates table_slice_builder
-void intrusive_ptr_add_ref(const table_slice_builder* ptr);
+// -- column builder helpers --------------------------------------------------
 
-/// @relates table_slice_builder
-void intrusive_ptr_release(const table_slice_builder* ptr);
+arrow::Status
+append_builder(const bool_type&, type_to_arrow_builder_t<bool_type>& builder,
+               const view<type_to_data_t<bool_type>>& view) noexcept;
+
+arrow::Status
+append_builder(const integer_type&,
+               type_to_arrow_builder_t<integer_type>& builder,
+               const view<type_to_data_t<integer_type>>& view) noexcept;
+
+arrow::Status
+append_builder(const count_type&, type_to_arrow_builder_t<count_type>& builder,
+               const view<type_to_data_t<count_type>>& view) noexcept;
+
+arrow::Status
+append_builder(const real_type&, type_to_arrow_builder_t<real_type>& builder,
+               const view<type_to_data_t<real_type>>& view) noexcept;
+
+arrow::Status
+append_builder(const duration_type&,
+               type_to_arrow_builder_t<duration_type>& builder,
+               const view<type_to_data_t<duration_type>>& view) noexcept;
+
+arrow::Status
+append_builder(const time_type&, type_to_arrow_builder_t<time_type>& builder,
+               const view<type_to_data_t<time_type>>& view) noexcept;
+
+arrow::Status
+append_builder(const string_type&,
+               type_to_arrow_builder_t<string_type>& builder,
+               const view<type_to_data_t<string_type>>& view) noexcept;
+
+arrow::Status
+append_builder(const pattern_type&,
+               type_to_arrow_builder_t<pattern_type>& builder,
+               const view<type_to_data_t<pattern_type>>& view) noexcept;
+
+arrow::Status
+append_builder(const address_type&,
+               type_to_arrow_builder_t<address_type>& builder,
+               const view<type_to_data_t<address_type>>& view) noexcept;
+
+arrow::Status
+append_builder(const subnet_type&,
+               type_to_arrow_builder_t<subnet_type>& builder,
+               const view<type_to_data_t<subnet_type>>& view) noexcept;
+
+arrow::Status
+append_builder(const enumeration_type&,
+               type_to_arrow_builder_t<enumeration_type>& builder,
+               const view<type_to_data_t<enumeration_type>>& view) noexcept;
+
+arrow::Status
+append_builder(const list_type& hint,
+               type_to_arrow_builder_t<list_type>& builder,
+               const view<type_to_data_t<list_type>>& view) noexcept;
+
+arrow::Status
+append_builder(const map_type& hint, type_to_arrow_builder_t<map_type>& builder,
+               const view<type_to_data_t<map_type>>& view) noexcept;
+
+arrow::Status
+append_builder(const record_type& hint,
+               type_to_arrow_builder_t<record_type>& builder,
+               const view<type_to_data_t<record_type>>& view) noexcept;
+
+template <type_or_concrete_type Type>
+arrow::Status
+append_builder(const Type& hint,
+               std::same_as<arrow::ArrayBuilder> auto& builder,
+               const std::same_as<data_view> auto& view) noexcept {
+  if (caf::holds_alternative<caf::none_t>(view))
+    return builder.AppendNull();
+  if constexpr (concrete_type<Type>) {
+    return append_builder(hint,
+                          caf::get<type_to_arrow_builder_t<Type>>(builder),
+                          caf::get<vast::view<type_to_data_t<Type>>>(view));
+  } else {
+    auto f
+      = [&]<concrete_type ResolvedType>(const ResolvedType& hint) noexcept {
+          return append_builder(
+            hint, caf::get<type_to_arrow_builder_t<ResolvedType>>(builder),
+            caf::get<vast::view<type_to_data_t<ResolvedType>>>(view));
+        };
+    return caf::visit(f, hint);
+  }
+}
 
 } // namespace vast
