+++ libvast/src/format/arrow.cpp
@@ -9,7 +9,6 @@
 #include "vast/format/arrow.hpp"
 
 #include "vast/arrow_table_slice.hpp"
-#include "vast/arrow_table_slice_builder.hpp"
 #include "vast/config.hpp"
 #include "vast/detail/assert.hpp"
 #include "vast/detail/byte_swap.hpp"
