+++ libvast/src/format/single_layout_reader.cpp
@@ -12,7 +12,6 @@
 #include "vast/factory.hpp"
 #include "vast/logger.hpp"
 #include "vast/table_slice_builder.hpp"
-#include "vast/table_slice_builder_factory.hpp"
 #include "vast/type.hpp"
 
 #include <chrono>
@@ -42,8 +41,7 @@ caf::error single_layout_reader::finish(consumer& f, caf::error result) {
 }
 
 bool single_layout_reader::reset_builder(type layout) {
-  VAST_TRACE_SCOPE("{} {}", VAST_ARG(table_slice_type_), VAST_ARG(layout));
-  builder_ = factory<table_slice_builder>::make(table_slice_type_, layout);
+  builder_ = std::make_shared<table_slice_builder>(std::move(layout));
   last_batch_sent_ = reader_clock::now();
   batch_events_ = 0;
   return builder_ != nullptr;
