+++ libvast/src/pipeline.cpp
@@ -10,7 +10,6 @@
 
 #include "vast/logger.hpp"
 #include "vast/table_slice_builder.hpp"
-#include "vast/table_slice_builder_factory.hpp"
 #include "vast/table_slice_encoding.hpp"
 
 #include <arrow/type.h>
