+++ libvast_test/src/events.cpp
@@ -18,7 +18,6 @@
 #include "vast/format/test.hpp"
 #include "vast/format/zeek.hpp"
 #include "vast/table_slice_builder.hpp"
-#include "vast/table_slice_builder_factory.hpp"
 #include "vast/type.hpp"
 
 #include <caf/settings.hpp>
@@ -44,8 +43,7 @@ struct alternating {};
 template <class Policy>
 std::vector<table_slice> make_integers(size_t count) {
   auto layout = type{"test.int", record_type{{"value", integer_type{}}}};
-  auto builder = factory<table_slice_builder>::make(
-    defaults::import::table_slice_type, layout);
+  auto builder = std::make_shared<table_slice_builder>(layout);
   VAST_ASSERT(builder != nullptr);
   std::vector<table_slice> result;
   result.reserve(count);
@@ -134,7 +132,6 @@ events::events() {
   static bool initialized = false;
   if (initialized)
     return;
-  factory<table_slice_builder>::initialize();
   initialized = true;
   // Read schemas
   std::ifstream base(artifacts::schemas::base);
