+++ plugins/parquet/tests/parquet.cpp
@@ -8,7 +8,6 @@
 
 #define SUITE parquet
 
-#include <vast/arrow_table_slice_builder.hpp>
 #include <vast/chunk.hpp>
 #include <vast/concept/parseable/to.hpp>
 #include <vast/concept/parseable/vast/expression.hpp>
@@ -23,6 +22,7 @@
 #include <vast/query_context.hpp>
 #include <vast/system/posix_filesystem.hpp>
 #include <vast/system/status.hpp>
+#include <vast/table_slice_builder.hpp>
 #include <vast/test/fixtures/actor_system_and_events.hpp>
 #include <vast/test/memory_filesystem.hpp>
 #include <vast/test/test.hpp>
@@ -36,7 +36,7 @@ namespace {
 template <class T, class... Ts>
 auto make_slice(const record_type& layout, std::vector<T> x0,
                 std::vector<Ts>... xs) {
-  auto builder = arrow_table_slice_builder::make(type{"rec", layout});
+  auto builder = std::make_shared<table_slice_builder>(type{"rec", layout});
   for (size_t i = 0; i < x0.size(); ++i) {
     CHECK(builder->add(x0.at(i)));
     if constexpr (sizeof...(Ts) > 0)
